Fixes an early release of a variable that's used in other blocks.
We discovered this during the UIGestureRecognize#recognize additions we were making. It seems that this wasn't a problem previously, but somehow our changes triggered an earlier release of the autorelease pool for the beforeEach block in this test. By retaining the window variable (and releasing in an afterEach block) rather than autoreleasing, this fixed the error.