Refactor Target fixture and use cedar spies
This seemed like a good idea, because when we initially made some changes to Target, we found that we were storing off a BOOL per method, and also potentially needed to store off a reference to the args that a method was called with. Since Cedar spies already provide us this functionality, it seemed like a bad idea to replicate it ourselves. Cedar spies are already used in a few other tests, so this doesn't seem like we're doing anything strange or unusual.