Refactor UIView gesture recognizer helper methods
@sweetleon and @akitchen both raised a good point that working with
blocks and dealing with stack/heap issues is a lot more of a headache
than just iterating over things cleanly. It seemed initially that
a block based approach would be cleaner, but looking at this diff,
it's pretty clear that this is a lot less code (and it's easier to
read, as well).