Allow `WKInterfaceObject`s with IBOutlets to exist in .storyboard files.
The InterfaceControllerLoader no longer throws a runtime exception when
it encounters an anonymous interface object in a storyboard.  This makes
its behavior more like production.

This means that storyboards under test can include labels, separators, etc.
that serve a purely cosmetic function and are never declared or addressed
in custom `WKInterfaceController` subclasses.  As such, these anonymous
interface objects are not currently exposed for writers of tests.