Avoid swizzling stateful properties on UIGestureRecognizers
If you were trying to observe the state of gesture recognizers using KVC,
or had expected any existing UIKit code that users the private interface
of gesture recognizers, then you would have a pretty bad time.

Since this state is already present in the implementation, it seems dangerous
to cover it up and reimplement it. The approach I've used here requires a
private API, but the UIKit SpecHelper target seems to be a reasonable place
for private API usage.