Add smoke test for Cocoapods installation for iOS
A new project is added to test one method from each of the following
subspecs:
  * FoundationCore
  * FoundationSpecHelper
  * UIKitCore
  * UIKitSpecHelper
  * CoreLocationSpecHelper

No assertions of the actual functionality are made just smoke tests to
ensure the frameworks are wired up correctly. The App uses
PivotalCoreKit/Core and the Spec uses PivotalCoreKit/Development to test
them independent of each other.