Rename the '-Dynamic' targets to be canonical.
Appends '-static' to the old static framework targets.
Adjusts Rake to refer to the old static framework.
Add Cedar to the header search path so that we only have to write
'#import "CDRSpecHelper.h"' in the cedar code.
Other small fixes to the code to get everything building working and
passing.

[finishes #107235232]