Watchkit classes respond to -reset_sent_messages
Useful if you need to call setHidden: on a watchkit UI element under test, but
cannot observer its hidden state with isHidden.

Say the initial state is hidden in the Storyboard, you may under test want to assert
that it was changed from visible back to hidden again. Without this, you won't be able to
assert that its final state is hidden without resetting the messages sent to the object.