-tap should raise when the control has no width or height
Encountered this on a project today. It felt very strange that PCK would allow
you to tap on controls that wouldn't be able to receive touch events in a real
setting. This sparked a discussion around "should we be using -clipsToBounds=Yes ?"
but decided that improving PCK to catch these errors earlier would be better.

For reference, the problem we needed to fix had to do with autolayout. We had forgot
to set some autolayout constraints correctly and ended up with a zero width, zero height
control that, while it looked okay, wasn't tappable.