UIAlertController spec helpers raise when not presented
The -dismiss spec helpers are intended to be used with UIAlertControllers
that have already been presented. For consistency, they make the presenting
view controller dismiss them, which is what the action would do in real life
when a user taps on these buttons.

However, when writing specs it's easy to forget to actually present them
and not understand why your tests are failing. This change should make it more
clear why the -dismiss actions do nothing at all when configured improperly.