Deprecate async NSURLConnection helpers
While it's hard to argue about the value of having fast tests, it seems dubious to make
inherently async things synchronous with a stub like this, especially since we've seen this
as the source of race conditions on CI lately.

We'll be marking this as deprecated immediately and will be removing it in a release VERY SOON.
If you're using this and are interested in continuing, I'd like to advocate that you fix the code
that sleeps on the current thread in this category to use a semaphore. Barring that, it's probably
best to migrate to a safer approach for verifying network requests, such as testing those
holistically without stubbing any of the behavior, without the rest of your app being tested as well.

Signed-off-by: Rachel Brindle <rbrindle@pivotal.io>